Wrote Description Import
The lines added within the Event Control loop are used for locating the
corresponding description file, importing/reading such file,
concatenating lines into one string (instead of an array), and
populating the text area with such string. However, the current version
of the code does not seem to recognize the setText() function. Also,
errors arise when additional buttons are pressed.