Button Type Verification System
In order to avid errors when pressing other buttons, a button-type
recognition process was created in several lines within the event
control loop. These lines use the name of the controller to identify the
button type. Here, I use the word type to differentiate buttons used for
scenarios vs buttons used for user id input, or other buttons. A
conditional statement is then used to verify that the button pressed is,
in fact, a button for scenarios.